fix the display of media conversions
